requirements.txt
fastapi
uvicorn
pydantic
aiogram
yt-dlp
ffmpeg-python
python-multipart
aiofiles
